feat: visibility image
