+++ watomate/Watomate/Resources/Assets.xcassets/Colors/yellow.colorset/Contents.json
@@ -5,8 +5,8 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "157",
-          "green" : "231",
+          "blue" : "80",
+          "green" : "217",
           "red" : "245"
         }
       },
@@ -23,8 +23,8 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "157",
-          "green" : "231",
+          "blue" : "80",
+          "green" : "217",
           "red" : "245"
         }
       },
