+++ watomate/Watomate/Resources/Assets.xcassets/Visibility/fl.imageset/Contents.json
@@ -1,52 +1,8 @@
 {
   "images" : [
     {
-      "filename" : "fl@1x.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "fl_dark@1x.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "fl@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "fl_dark@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "filename" : "fl@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "fl_dark@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
+      "filename" : "팔로워공개 (2).svg",
+      "idiom" : "universal"
     }
   ],
   "info" : {
