+++ watomate/Watomate/Resources/Assets.xcassets/Visibility/pb.imageset/Contents.json
@@ -1,52 +1,8 @@
 {
   "images" : [
     {
-      "filename" : "pb@1x.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "pb_dark@1x.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "pb@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "pb_dark@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "filename" : "pb@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "pb_dark@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
+      "filename" : "전체공개 (1).svg",
+      "idiom" : "universal"
     }
   ],
   "info" : {
