+++ watomate/Watomate/Resources/Assets.xcassets/Visibility/pr.imageset/Contents.json
@@ -1,52 +1,8 @@
 {
   "images" : [
     {
-      "filename" : "pr@1x.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "pr_dark@1x.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "pr@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "pr_dark@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "filename" : "pr@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "filename" : "pr_dark@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
+      "filename" : "나만보기 (1).svg",
+      "idiom" : "universal"
     }
   ],
   "info" : {
